+++ CHANGELOG.md
@@ -245,6 +245,7 @@
 - Entities: auto create dir **[DONE]**
 - Rust server API documentation **[DONE]**
 - Object API: correct nested `set` usage for request **[DONE]**
+- Object API: add nested `upsert` for single relation **[DONE]**
 - Object API: correct nested `set` usage for programming **[DONE]**
 - `find` action transform when including
 - Test soft delete support